% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDoc.R
\docType{package}
\name{trdist}
\alias{trdist-package}
\alias{trdist}
\title{Univariate Proability Distributions with Truncation}
\description{
Truncation of univariate probability distributions. The probability distribution can come from other packages so long as the function names follow the standard d, p, q, r naming format. Also other univariate probability distributions are included.
}
\author{
\strong{Maintainer}: Jared Studyvin \email{studyvinstat@gmail.com}

}
\keyword{package}
